refactor(authority): move v2 signing crypto into @stwd/shared (proxy verifier reach)
Contradiction C2 (reported): the proxy runs as a SEPARATE process and does not
depend on @stwd/api, but the governed claim must verify the v2 commitment. Move
the pure v2 signing crypto (HKDF keyId rotation, sign/verify, fail-closed) from
the api service into @stwd/shared/provider-execution-auth (a proxy dependency).
api re-exports the symbols for existing callers. 8/8 crypto tests still pass.
