+++ packages/shared/src/index.ts
@@ -10,6 +10,7 @@ export * from "./execution-payload.js";
 export type { PriceOracle } from "./price-oracle.js";
 export { createPriceOracle } from "./price-oracle.js";
 export * from "./provider-action.js";
+export * from "./provider-execution-auth.js";
 export * from "./provider-approval.js";
 export * from "./provider-authority.js";
 // ─── Non-throwing description of an arbitrary thrown value (fail-closed catches) ───
